+++ src/data/ActiveDataProvider.php
@@ -2,6 +2,11 @@
 
 namespace davidhirtz\yii2\skeleton\data;
 
+use davidhirtz\yii2\skeleton\db\ActiveRecord;
+
+/**
+ * @template T of ActiveRecord
+ */
 class ActiveDataProvider extends \yii\data\ActiveDataProvider
 {
     public const EVENT_INIT = 'init';
@@ -15,8 +20,16 @@ class ActiveDataProvider extends \yii\data\ActiveDataProvider
         $this->prepareQuery();
     }
 
-    public function prepareQuery(): void
+    protected function prepareQuery(): void
     {
         $this->trigger(static::EVENT_AFTER_PREPARE);
+    }
+
+    /**
+     * @return T[]
+     */
+    public function getModels(): array
+    {
+        return parent::getModels();
     }
 }